feat(go): propagate traceparents from parents
